fix: use a instead of buttons and rebase fixups
