+++ frontend/static/auth.html
@@ -3,12 +3,12 @@
 
     <p class="mb-4">Choose a method to authenticate:</p>
     <div class="d-grid gap-2 col-6 mx-auto">
-        <button id="loginButton" class="btn btn-primary btn-lg" onclick="router.navigate('/login')">Login</button>
-        <button id="login42Button" class="btn btn-success btn-lg" onclick="router.navigate('/login/42')">Login with 42</button>
+        <a id="loginButton" class="btn btn-primary btn-lg" href="/login">Login</a>
+        <a id="login42Button" class="btn btn-success btn-lg" href="/login/42">Login with 42</a>
     </div>
 
     <p class="mt-4">Or create an account:</p>
     <div class="d-grid gap-2 col-6 mx-auto">
-        <button id="registerButton" class="btn btn-secondary btn-lg" onclick="router.navigate('/register')">Register</button>
+        <a id="registerButton" class="btn btn-secondary btn-lg" href="/register">Register</a>
     </div>
 </div>
