--- frontend/static/bonjour.html
@@ -1,1 +1,0 @@
-<p>Ceci est la page secondaire appelée "bonjour"</p>
