+++ frontend/static/home.html
@@ -8,7 +8,7 @@
     <div class="mt-5">
         <h2>Play Pong</h2>
         <p>Challenge your friends in this modern version of the iconic Pong game.</p>
-        <button class="btn btn-primary" onclick="router.navigate('/pong')">Start Playing Pong</button>
+        <a class="btn btn-primary" href="/pong">Start Playing Pong</a>
     </div>
 
 
@@ -16,6 +16,6 @@
     <div class="mt-5">
         <h3>Ready to Transcend?</h3>
         <p>Sign in and start your gaming journey today!</p>
-        <button class="btn btn-success" onclick="router.navigate('/sign-in')">Sign In Now</button>
+        <a class="btn btn-success" href="/sign-in">Sign In Now</a>
     </div>
 </div>
