+++ frontend/static/index.html
@@ -12,20 +12,20 @@
     <!-- Navbar -->
     <nav class="navbar navbar-expand-lg fixed-top">
         <div class="container-fluid">
-            <a class="navbar-brand" href="#" onclick="router.navigate('/'); return false;">Transcendence</a>
+            <a class="navbar-brand" href="/">Transcendence</a>
             <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
                 <span class="navbar-toggler-icon"></span>
             </button>
             <div class="collapse navbar-collapse" id="navbarNav">
                 <ul class="navbar-nav ms-auto">
                     <li class="nav-item">
-                        <button class="btn btn-link nav-link" onclick="router.navigate('/sign-in')">Sign In</button>
+                        <a class="btn btn-link nav-link" href="/sign-in">Sign In</a>
                     </li>
                     <li class="nav-item">
-                        <button class="btn btn-link nav-link" onclick="router.navigate('/logout')">Login with 42</button>
+                        <a class="btn btn-link nav-link" href="/login/42">Login with 42</a>
                     </li>
                     <li class="nav-item">
-                        <button class="btn btn-link nav-link" onclick="router.navigate('/users')">Login with 42</button>
+                        <a class="btn btn-link nav-link" href="/users">Users</a>
                     </li>
                 </ul>
             </div>
